Queue length 1 modifications
These modifications change the behavior of Fictrac's frame queueing system so that only a single frame can be buffered.  Should the queue increase past a single element only the most recent will be stored and used in computation.  Should such an event occur, a counter (displayed in the cmd line) will be updated.